requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
librosa==0.10.1
numpy==1.24.3
tensorflow==2.13.0
llama-cpp-python==0.2.11
python-dotenv==1.0.0
soundfile>=0.12.1
matplotlib>=3.7.0
scipy>=1.7.0
requests>=2.26.0 